# swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Txing AI API文档",
        "title": "Txing AI API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/channel": {
            "post": {
                "description": "创建新的渠道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "渠道管理"
                ],
                "summary": "创建渠道",
                "parameters": [
                    {
                        "description": "渠道信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.ChannelVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/channel/list": {
            "get": {
                "description": "获取渠道列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "渠道管理"
                ],
                "summary": "获取渠道列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式(asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "渠道类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "渠道名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/channel/{id}": {
            "get": {
                "description": "获取指定渠道的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "渠道管理"
                ],
                "summary": "获取渠道详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.ChannelVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新渠道信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "渠道管理"
                ],
                "summary": "更新渠道",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "渠道信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.ChannelVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定渠道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "渠道管理"
                ],
                "summary": "删除渠道",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/api/cos/presigned-url": {
            "post": {
                "description": "获取文件上传或下载的预签名URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对象存储相关"
                ],
                "summary": "获取预签名URL",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetPresignedURLReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.GetPresignedURLVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateChannelReq": {
            "type": "object",
            "required": [
                "endpoint",
                "models",
                "name",
                "secret",
                "type"
            ],
            "properties": {
                "endpoint": {
                    "description": "服务地址 Service endpoint",
                    "type": "string"
                },
                "mappings": {
                    "description": "模型映射关系 Mappings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/global.ModelMapping"
                    }
                },
                "models": {
                    "description": "支持的模型列表 Supported models",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "渠道名称 Channel name",
                    "type": "string"
                },
                "priority": {
                    "description": "优先级 Priority",
                    "type": "integer"
                },
                "retry": {
                    "description": "重试次数 Retry times",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥 Secret key",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 Enable status",
                    "type": "boolean"
                },
                "type": {
                    "description": "渠道类型 Channel type",
                    "type": "string"
                },
                "weight": {
                    "description": "权重 Weight",
                    "type": "integer"
                }
            }
        },
...
        "dto.GetPresignedURLReq": {
            "type": "object",
            "required": [
                "key",
                "type"
            ],
            "properties": {
                "key": {
                    "description": "对象键",
                    "type": "string"
                },
                "type": {
                    "description": "类型：upload/download",
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateChannelReq": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "description": "服务地址 Service endpoint",
                    "type": "string"
                },
                "mappings": {
                    "description": "模型映射关系 Mappings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/global.ModelMapping"
                    }
                },
                "models": {
                    "description": "支持的模型列表 Supported models",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "渠道名称 Channel name",
                    "type": "string"
                },
                "priority": {
                    "description": "优先级 Priority",
                    "type": "integer"
                },
                "retry": {
                    "description": "重试次数 Retry times",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥 Secret key",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 Enable status",
                    "type": "boolean"
                },
                "type": {
                    "description": "渠道类型 Channel type",
                    "type": "string"
                },
                "weight": {
                    "description": "权重 Weight",
                    "type": "integer"
                }
            }
        },
...
        "global.ModelMapping": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "条件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/global.ModelMappingCondition"
                    }
                },
                "sourceModel": {
                    "description": "源模型",
                    "type": "string"
                }
            }
        },
        "global.ModelMappingCondition": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "条件映射，key 为条件名称，value 为条件值",
                    "type": "object",
                    "additionalProperties": true
                },
                "targetModel": {
                    "description": "目标模型",
                    "type": "string"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "vo.ChannelVO": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建时间 Create time",
                    "type": "string"
                },
                "endpoint": {
                    "description": "服务地址 Service endpoint",
                    "type": "string"
                },
                "id": {
                    "description": "渠道ID Channel ID",
                    "type": "integer"
                },
                "mappings": {
                    "description": "模型映射关系 Mappings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/global.ModelMapping"
                    }
                },
                "models": {
                    "description": "支持的模型列表 Supported models",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "渠道名称 Channel name",
                    "type": "string"
                },
                "priority": {
                    "description": "优先级 Priority",
                    "type": "integer"
                },
                "retry": {
                    "description": "重试次数 Retry times",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥 Secret key",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 Enable status",
                    "type": "boolean"
                },
                "type": {
                    "description": "渠道类型 Channel type",
                    "type": "string"
                },
                "update_at": {
                    "description": "更新时间 Update time",
                    "type": "string"
                },
                "weight": {
                    "description": "权重 Weight",
                    "type": "integer"
                }
            }
        },
...
        "vo.GetPresignedURLVO": {
            "type": "object",
            "properties": {
                "expireAt": {
                    "description": "过期时间戳",
                    "type": "integer"
                },
                "key": {
                    "description": "对象键",
                    "type": "string"
                },
                "url": {
                    "description": "预签名URL",
                    "type": "string"
                }
            }
        },
...
    }
}